Item 6: Manager Config value validation

+++ contracts/croncat-manager/src/helpers.rs
@@ -249,7 +249,7 @@ pub(crate) fn finalize_task(
     // Sub native for gas
     let gas_with_fees = gas_with_fees(
         queue_item.task.amount_for_one_task.gas,
-        config.agent_fee + config.treasury_fee,
+        (config.agent_fee + config.treasury_fee) as u64,
     )?;
     let native_for_gas_required = config.gas_price.calculate(gas_with_fees).unwrap();
     task_balance.native_balance = task_balance
